Refactor G4HepEmData Json I/O into dedicated library
To keep G4HepEmData as lighweight as possible, separate Json
serialization code and interfaces to dedicated "g4HepEmDataJsonIO" library.
This also allows future serialization options free, and can just add
further libraries as required/optionally (e.g. HDF5/ROOT)

Update tests to use new library.
