+++ G4HepEm/CMakeLists.txt
@@ -1,5 +1,6 @@
 # - Build component libraries
 add_subdirectory(G4HepEmData)
+add_subdirectory(g4HepEmDataJsonIO)
 add_subdirectory(G4HepEmInit)
 add_subdirectory(G4HepEmRun)
