+++ G4HepEm/G4HepEmData/CMakeLists.txt
@@ -15,11 +15,9 @@ target_compile_features(g4HepEmData PUBLIC cxx_std_${CMAKE_CXX_STANDARD})
 
 target_include_directories(g4HepEmData PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-  $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}/${PROJECT_NAME}>
-  PRIVATE
-  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src>)
+  $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}/${PROJECT_NAME}>)
 
-# TODO: Determine when/if to use CUDA::cudart_static
+ # TODO: Determine when/if to use CUDA::cudart_static
 target_link_libraries(g4HepEmData PRIVATE $<$<BOOL:${G4HepEm_CUDA_BUILD}>:CUDA::cudart>)
 
 ## ----------------------------------------------------------------------------
