+++ testing/DataImportExport/CMakeLists.txt
@@ -7,7 +7,7 @@ target_include_directories(TestDataImportExport PRIVATE ${CMAKE_CURRENT_SOURCE_D
 
 target_link_libraries(TestDataImportExport
   PRIVATE
-  g4HepEm TestUtils ${Geant4_LIBRARIES})
+  g4HepEm g4HepEmDataJsonIO TestUtils ${Geant4_LIBRARIES})
 
 add_test(NAME TestDataImportExportFull COMMAND TestDataImportExport full)
 add_test(NAME TestDataImportExportSimple COMMAND TestDataImportExport simple)
